Fix for sink policy PD data transfer state machine issue
To fix the sink port policy engine PD data transfer timer expiry
issue, optimized the data transferring path from protocol to sink port
policy engine. Also reinitialized the completion data structure for
avoiding the state machine error and handled the sink port PE states
appropriately.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-37820

+++ drivers/usb/typec/pd/protocol.h
@@ -41,7 +41,7 @@ struct pd_prot {
 	u8 assumed_pwr_role;
 	u8 event;
 	u8 tx_msg_id;
-	s8 rx_msg_id;
+	s32 rx_msg_id;
 	u8 retry_counter;
 
 	u8 cur_tx_state;
